requirements.txt
streamlit
Pillow
sentence-transformers
langchain
transformers
torch
chromadb
scikit-learn
pandas
plotly
umap-learn
